refactor: move RouteAnnouncer to seperate file
